Feat: fixed product creation module
Fixed file upload on product creation

+++ Controllers/ProductController.cs
@@ -1,6 +1,5 @@
 using System.Security.Claims;
 using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http.Features;
 using Microsoft.AspNetCore.Mvc;
 using TheStore.Models;
 using TheStore.Models.Users;
@@ -12,65 +11,35 @@ namespace TheStore.Controllers;
 [Route("api/products")]
 public class ProductController : ControllerBase {
   private readonly ProductService productService;
-  private readonly UploadService uploadService;
   private readonly RepoService repo;
-  public ProductController(ProductService _productService, RepoService _repo, UploadService _uploadService) {
+  public ProductController(ProductService _productService, RepoService _repo) {
     productService = _productService;
     repo = _repo;
-    uploadService = _uploadService;
   }
 
-  
   [HttpPost("new", Name = "NewProduct"), Authorize]
   [ProducesResponseType(StatusCodes.Status200OK)]
   [ProducesResponseType(StatusCodes.Status401Unauthorized)]
   [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-  public ActionResult NewProduct([FromBody] ProductDTO productDto){
+  public ActionResult NewProduct([FromForm] ProductDTO productDto, List<IFormFile> Pictures){
     UserClaims? user = GetCurrentUser();
-    var response = productService.NewProduct(productDto, Guid.Parse(user!.UserId));
+    if(user == null) return Unauthorized();
+    if(user.UserId == null) return Unauthorized();
+    var response = productService.NewProduct(productDto, Guid.Parse(user.UserId), Pictures);
     if(!response.Success) return BadRequest(response);
     return Ok(response);
   }
 
 
-  [HttpGet("admin/list", Name = "GetAdminProducts")]
+  [HttpGet("admin-products", Name = "GetAdminProducts")]
   [ProducesResponseType(StatusCodes.Status200OK)]
   [ProducesResponseType(StatusCodes.Status401Unauthorized)]
   [ProducesResponseType(StatusCodes.Status400BadRequest)]
   [ProducesResponseType(StatusCodes.Status500InternalServerError)]
   public ActionResult GetAdminProducts(){
-    // var response = productService.GetAdminProducts();
-    // if(!response.Success) return BadRequest(response);
-    var res = (
-      from product in repo.Products
-      join brand in repo.Brands
-      on product.BrandId equals brand.BrandId
-      join category in repo.Categories
-      on product.CategoryId equals category.CategoryId
-      join stock in repo.Stocks
-      on product.Id equals stock.ProductId
-      select new {
-        product.Id,
-        product.ProductName,
-        product.ProductDescription,
-        product.Price,
-        stock = new {
-          stock.StockId,
-          stock.Quantity,
-          stock.ReorderLevel,
-          stock.CostPrice
-        },
-        brand = new {
-          brand.BrandId,
-          brand.BrandName
-        },
-        category = new {
-          category.CategoryId,
-          category.CategoryName
-        }
-      }
-    ).ToList();
-    return Ok(res);
+    var response = productService.GetAdminProducts();
+    if(!response.Success) return BadRequest(response);
+    return Ok(response);
   }
 
   
@@ -112,21 +81,13 @@ public class ProductController : ControllerBase {
   [ProducesResponseType(StatusCodes.Status500InternalServerError)]
   public ActionResult DeleteProduct(Guid Id) {
     UserClaims? user = GetCurrentUser();
+    if(user == null) return Unauthorized();
+    if(user.UserId == null) return Unauthorized();
     var response = productService.DeleteProduct(Id, Guid.Parse(user.UserId));
     if(!response.Success) return BadRequest(response);
     return Ok(response);
   }
 
-
-  [HttpPost("file/upload/{productId}")]
-  [ProducesResponseType(StatusCodes.Status200OK)]
-  [ProducesResponseType(StatusCodes.Status400BadRequest)]
-  [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-  public ActionResult UploadPicture(List<IFormFile> files, Guid productId) {
-    var res = uploadService.UploadPictures(files, productId);
-    if(!res.Success) return BadRequest(res);
-    return Ok(res);
-  }
 
   // get current user from token
   private UserClaims? GetCurrentUser(){
